Add vertical sweep
